RHUI: drop /content/fastrack/rhel/rhui alias
I've gone through and checked all RHUI aliases against actual
data on CDN. This one seems to be incorrect:

1. on current CDN filesystem, the rhui path doesn't exist at all
2. on older filesystem, it exists, but is not symlinked to non-rhui
3. on older filesystem, content between rhui and non-rhui paths
   doesn't appear to be identical
4. the relevant content sets were disabled in 2014 (rcm-metadata
   ec3025f411e46c)

Hence this is not actually a valid RHUI alias.  It was probably
intended to be, so an argument could be made for keeping it, but
it would force us to change point (3) above and start serving
identical content for the two paths, and there's not really any
benefit to making such a change now, only risk.