Create cdn_definitions.api module [RHELDST-4862]
The cdn_definitions.api module allows clients to consume the
data.json or the schema.json files via a convenient API.

This commit also introduces an optional "source" argument to the
data_loads method. The source argument allows the user to specify
the source of the cdn_definitions data.json file. The source could
be a URL or a local path on a directory tree. If a source is not
specified, the location may be overriden via the
CDN_DEFINITIONS_PATH environment variable, which also may either
be in the form of a URL or a local path.