Drop ineffective requiredProperties
The schema here attempted to require certain properties, but this had no
effect because the correct keyword is "required", not
"requiredProperties". This can be confirmed at
http://json-schema.org/draft-07/schema.

So, in reality we have never required any properties here.
We could now change this to "required" to match the original intent,
but I think it's reasonable to just drop them instead since:

- adding new required properties is backwards-incompatible. In the
  worst-case scenario it breaks some services; in the best case it
  probably breaks at least some tests.

- it anyway seems more user-friendly to allow these fields to be
  omitted, since they always have some obvious default value
  which could be provided (such as an empty list or dict). These could
  even be filled in by the library automatically on load, though this
  is not pursued at the moment.