Add tests to verify current schema is backward compatible
All schema changes ought to be backwards compatible, as any
usage would be prone to break any upgrade to cdn-definitions.
Hence, add the test to validate historical data against current
schema.
The historical data in tests/historical-data are the data.yamls
that were present preceding any schema updates i.e. data-1.0.0.yaml
has that data that was present when the schema was updated in
v2.0.0., so the test could validate the old data against this
updated schema.