Convert RHUI product ID to list of IDs [RHELDST-28614]
Recently created rhui tooling products and content sets caused issues for downstream tools,
such as Manifest-API.  This is because their download_urls contained "/rhui", which
erroneously set the for_rhui property to true.

This issue has been fixed, but we would like to extend the solution to
make it more robust. This change adds a new rhui_product_ids entry to
the schema, so we can compare against a list of IDs rather than a single
one. This was done to avoid breaking changes.