Make repo content sync configurable [RHELDST-34407]
Currently, the configuration of the source repositories for content
sync from RHEL to UBI is hardcoded in cdn-utils, and the same goes
also for Client Tools. This makes any addition of more synced content
difficult, as it requires tooling changes and deployment each time.

Additionally, `populate_ubi_dot_repos` isn't flexible enough because
it doesn't allow for selective control over which products have their
content populated in the DOT repos and in which environments.

This commit adds `repo_content_sync` configuration that allows to set
the content synchronization per-environment. It accepts a list of git
repo paths containing the configuration for tooling that facilitates
the repo content sync (currently ubipop) and a flag that controls
DOT repo population per-source.